+++ lib/models/list_item.dart
